Clean up Abstract_Class_Parser befpre looking at issue 525
+ For class `c_string_vector`, move member fuction definitions into the `.i.hh` file. This allows
  the entire helper class to be defined in a single place.  Also disable copy/move constructors and
  assignment/move-assignment operations.
+ For class `Contextual_Parse_Functor`, move all definitions into to a new `.hh` file. This allows
  the entire helper class to be defined in a single place.
+ Re-wrap comment blocks to 100 columns.
+ Remove doxygen config option that is deprecated by newer versions of doxygen.
