+++ src/parser/Abstract_Class_Parser.i.hh
@@ -4,7 +4,7 @@
  * \author Kent Budge
  * \date   Thu Jul 17 14:08:42 2008
  * \brief  Member definitions of class Abstract_Class_Parser
- * \note   Copyright (C) 2016-2020 Triad National Security, LLC., All rights reserved */
+ * \note   Copyright (C) 2010-2021 Triad National Security, LLC., All rights reserved. */
 //------------------------------------------------------------------------------------------------//
 
 #ifndef utils_Abstract_Class_Parser_i_hh
@@ -13,36 +13,39 @@
 //! \bug This file needs to be cleaned up for doxygen parsing.
 //! \cond doxygen_ignore_block
 
-//------------------------------------------------------------------------------------------------//
-template <typename Abstract_Class, typename Context, Context const &get_context()>
-Contextual_Parse_Functor<Abstract_Class, Context, get_context>::Contextual_Parse_Functor(
-    std::shared_ptr<Abstract_Class> parse_function(Token_Stream &, Context const &))
-    : f_(parse_function) {}
-
-template <typename Abstract_Class, typename Context, Context const &get_context()>
-std::shared_ptr<Abstract_Class> Contextual_Parse_Functor<Abstract_Class, Context, get_context>::
-operator()(Token_Stream &tokens) const {
-  return f_(tokens, get_context());
-}
-
 //================================================================================================//
 /*!
- * Helper class defining a table of raw strings created by strdup that will be
- * properly deallocated using free on program termination.
+ * \class c_string_vector
+ * \brief Helper class defining a table of raw strings created by strdup that will be properly
+ *        deallocated using \c delete[] on program termination.
  */
 class c_string_vector {
 public:
-  ~c_string_vector();
-  c_string_vector() : data(0) { /* empty */
+  //! destructor
+  ~c_string_vector() {
+    for (auto &d : data)
+      delete[] d;
   }
+  //! constructor
+  c_string_vector() : data(0) {}
+  // Disallow copy/move/assignment/move-assignment operators.
+  c_string_vector(const c_string_vector &) = delete;
+  c_string_vector(const c_string_vector &&) = delete;
+  c_string_vector operator=(const c_string_vector &) = delete;
+  c_string_vector operator=(const c_string_vector &&) = delete;
+
+  // DATA
+
+  //! The tabe of raw strings
   std::vector<char *> data;
 };
+
+//! Keep this data as a global static variable.
 DLL_PUBLIC_parser extern c_string_vector abstract_class_parser_keys;
 
 //================================================================================================//
 /*
- * The following rather lengthy and clumsy declaration declares storage for the
- * parse functions.
+ * The following rather lengthy and clumsy declaration declares storage for the parse functions.
  *
  * Remember:
  * \code
@@ -56,13 +59,12 @@ std::vector<Parse_Function>
 
 //------------------------------------------------------------------------------------------------//
 /*!
- * This function allows a host code to register children of the abstract class
- * with the parser. This helps support extensions by local developers.
+ * \brief This function allows a host code to register children of the abstract class with the
+ *        parser. This helps support extensions by local developers.
  *
- * \param keyword Keyword associated with the child class
- *
- * \param parsefunction Parse function that reads a specification from a
- *           Token_Stream and returns a corresponding object of the child class.
+ * \param[in] keyword Keyword associated with the child class
+ * \param parsefunction Parse function that reads a specification from a Token_Stream and returns a
+ *               corresponding object of the child class.
  */
 template <typename Class, Parse_Table &get_parse_table(),
           std::shared_ptr<Class> &get_parsed_object(), typename Parse_Function>
@@ -88,13 +90,12 @@ void Abstract_Class_Parser<Class, get_parse_table, get_parsed_object,
 
 //------------------------------------------------------------------------------------------------//
 /*!
- * This function allows a host code to register children of the abstract class
- * with the parser. This helps support extensions by local developers.
+ * This function allows a host code to register children of the abstract class with the parser. This
+ * helps support extensions by local developers.
  *
- * \param keyword Keyword associated with the child class
- *
- * \param parsefunction Parse function that reads a specification from a
- *           Token_Stream and returns a corresponding object of the child class.
+ * \param[in] keyword Keyword associated with the child class
+ * \param parsefunction Parse function that reads a specification from a Token_Stream and returns a
+ *           corresponding object of the child class.
  */
 template <typename Class, Parse_Table &get_parse_table(),
           std::shared_ptr<Class> &get_parsed_object(), typename Parse_Function>
@@ -122,8 +123,8 @@ void Abstract_Class_Parser<Class, get_parse_table, get_parsed_object,
 
 //------------------------------------------------------------------------------------------------//
 /*!
- * This is the generic parse function associated with all child keywords. It
- * makes use of the Parse_Function associated with each child keyword.
+ * This is the generic parse function associated with all child keywords. It makes use of the
+ * Parse_Function associated with each child keyword.
  */
 template <typename Class, Parse_Table &get_parse_table(),
           std::shared_ptr<Class> &get_parsed_object(), typename Parse_Function>
